[CMLG-002] fix the bug in master branch when converting data from excel to database
Issue:

In TranslationImport file, the code is looking for language with value "zh_cn" while in excel the corresponding language name is 'cn' only.

Excel can contain complete empty rows which will also be read but shouldn't be put into the database.

The index of the data stored in the database sometimes doesn't start with 1.

The /import endpoint shouldn't be exposed to the public.

Solution:

Changing the language used in TranslationImport from 'zh_cn' to 'cn' to match the excel sheet.
Adding functions to check if the whole row is completely empty, and only process that row if it contains some data.
Determining the id of each model when inserting it into the database.
Comment out the /import endpoint

Risk:
Developers who want to set up a local database themselves need to uncomment the /import endpoint.
Calling the /import endpoint twice will cause an exception being thrown. To reset your database, delete everything in your database first and then call the /import endpoint.

Reviewed by:
Eileen, Annie